refactor : HandCard 클래스의 소속 패키지와 메서드 변경
- 카드의 집합을 다루는 클래스이기 때문에 user 패키지보다는 card 패키지에 어울린다고 판단함
- 보관 중인 카드 이름들을 나열한 문자열을 만드는 메서드는 view의 기능이기 때문에 제거
